fix(auth): change @Transactional scope to fix not increasing number of attempts in wrong OTP when user register and remove unused JWT methods
